+++ HISTORY.md
@@ -9,6 +9,9 @@ add upcoming unreleased modifications here
 * Breaking changes
   * commandline interface changed. See cli help via `nichtparasoup --help`.
 
+* Changes
+  * `nichtparasoup.testing.config.ConfigFileTest.probe()` got an optional second param `delay`.
+
 ## 2.4.1
 
 * Fixed